REFACTOR STUFF ...!
